+++ pyluxa4/server.py
@@ -2,14 +2,21 @@
 from flask import Flask, jsonify, abort, make_response, request
 from flask_httpauth import HTTPTokenAuth
 from gevent.pywsgi import WSGIServer
+from gevent.lock import BoundedSemaphore
+import gevent
+import os
+from . import scheduler
 from . import usb
 from . import common as cmn
 from . import __meta__
+
+sem = BoundedSemaphore(1)
 
 app = Flask(__name__)
 auth = HTTPTokenAuth('Bearer')
 tokens = set()
 luxafor = None
+schedule = None
 HOST = '0.0.0.0'
 PORT = 5000
 
@@ -18,27 +25,6 @@ def get_api_ver_path():
     """Get the API path."""
 
     return '/pyluxa4/api/v%s.%s' % __meta__.__version_info__[:2]
-
-
-def is_bool(name, value):
-    """Check if bool."""
-
-    if not isinstance(value, bool):
-        raise TypeError("'{}' must be a boolean".format(name))
-
-
-def is_str(name, value):
-    """Check if bool."""
-
-    if not isinstance(value, str):
-        raise TypeError("'{}' must be a string".format(name))
-
-
-def is_int(name, value):
-    """Check if bool."""
-
-    if not isinstance(value, int):
-        raise TypeError("'{}' must be a integer".format(name))
 
 
 @auth.verify_token
@@ -56,15 +42,17 @@ def color():
     try:
         error = ''
         led = request.json.get("led", cmn.LED_ALL)
-        is_int('led', led)
+        cmn.is_int('led', led)
         color = request.json.get('color', '')
-        is_str('color', color)
+        cmn.is_str('color', color)
     except Exception as e:
         error = str(e)
 
     if not error:
         try:
+            sem.acquire()
             luxafor.color(color, led=led)
+            sem.release()
         except Exception as e:
             error = str(e)
 
@@ -74,7 +62,7 @@ def color():
     return jsonify(
         {
             "path": request.path,
-            "status": 'success' if not error else 'fail',
+            "status": 'success',
             "code": 200,
             "error": error
         }
@@ -87,19 +75,19 @@ def fade():
     try:
         error = ''
         led = request.json.get("led", cmn.LED_ALL)
-        is_int('led', led)
+        cmn.is_int('led', led)
         color = request.json.get('color', '')
-        is_str('color', color)
+        cmn.is_str('color', color)
         speed = request.json.get('speed', 0)
-        is_int('speed', speed)
-        wait = request.json.get('wait', False)
-        is_bool('wait', wait)
+        cmn.is_int('speed', speed)
     except Exception as e:
         error = str(e)
 
     if not error:
         try:
-            luxafor.fade(color, led=led, speed=speed, wait=wait)
+            sem.acquire()
+            luxafor.fade(color, led=led, speed=speed)
+            sem.release()
         except Exception as e:
             error = str(e)
 
@@ -109,7 +97,7 @@ def fade():
     return jsonify(
         {
             "path": request.path,
-            "status": 'success' if not error else 'fail',
+            "status": 'success',
             "code": 200,
             "error": error
         }
@@ -122,21 +110,21 @@ def strobe():
     try:
         error = ''
         led = request.json.get("led", cmn.LED_ALL)
-        is_int('led', led)
+        cmn.is_int('led', led)
         color = request.json.get('color', '')
-        is_str('color', color)
+        cmn.is_str('color', color)
         speed = request.json.get('speed', 0)
-        is_int('speed', speed)
+        cmn.is_int('speed', speed)
         repeat = request.json.get('repeat', 0)
-        is_int('repeat', repeat)
-        wait = request.json.get('wait', False)
-        is_bool('wait', wait)
+        cmn.is_int('repeat', repeat)
     except Exception as e:
         error = str(e)
 
     if not error:
         try:
-            luxafor.strobe(color, led=led, speed=speed, repeat=repeat, wait=wait)
+            sem.acquire()
+            luxafor.strobe(color, led=led, speed=speed, repeat=repeat)
+            sem.release()
         except Exception as e:
             error = str(e)
 
@@ -146,7 +134,7 @@ def strobe():
     return jsonify(
         {
             "path": request.path,
-            "status": 'success' if not error else 'fail',
+            "status": 'success',
             "code": 200,
             "error": error
         }
@@ -159,21 +147,21 @@ def wave():
     try:
         error = ''
         color = request.json.get('color', '')
-        is_str('color', color)
+        cmn.is_str('color', color)
         wave = request.json.get('wave', cmn.WAVE_SHORT)
-        is_int('wave', wave)
+        cmn.is_int('wave', wave)
         speed = request.json.get('speed', 0)
-        is_int('speed', speed)
+        cmn.is_int('speed', speed)
         repeat = request.json.get('repeat', 0)
-        is_int('repeat', repeat)
-        wait = request.json.get('wait', False)
-        is_bool('wait', wait)
+        cmn.is_int('repeat', repeat)
     except Exception as e:
         error = str(e)
 
     if not error:
         try:
-            luxafor.wave(color, wave=wave, speed=speed, repeat=repeat, wait=wait)
+            sem.acquire()
+            luxafor.wave(color, wave=wave, speed=speed, repeat=repeat)
+            sem.release()
         except Exception as e:
             error = str(e)
 
@@ -183,7 +171,7 @@ def wave():
     return jsonify(
         {
             "path": request.path,
-            "status": 'success' if not error else 'fail',
+            "status": 'success',
             "code": 200,
             "error": error
         }
@@ -196,17 +184,17 @@ def pattern():
     try:
         error = ''
         pattern = request.json.get('pattern', 0)
-        is_int('pattern', pattern)
+        cmn.is_int('pattern', pattern)
         repeat = request.json.get('repeat', 0)
-        is_int('repeat', repeat)
-        wait = request.json.get('wait', False)
-        is_bool('wait', wait)
+        cmn.is_int('repeat', repeat)
     except Exception as e:
         error = str(e)
 
     if not error:
         try:
-            luxafor.pattern(pattern, repeat=repeat, wait=wait)
+            sem.acquire()
+            luxafor.pattern(pattern, repeat=repeat)
+            sem.release()
         except Exception as e:
             error = str(e)
 
@@ -216,7 +204,7 @@ def pattern():
     return jsonify(
         {
             "path": request.path,
-            "status": 'success' if not error else 'fail',
+            "status": 'success',
             "code": 200,
             "error": error
         }
@@ -228,7 +216,9 @@ def off():
 
     error = ''
     try:
+        sem.acquire()
         luxafor.off()
+        sem.release()
     except Exception as e:
         error = str(e)
 
@@ -238,7 +228,7 @@ def off():
     return jsonify(
         {
             "path": request.path,
-            "status": 'success' if not error else 'fail',
+            "status": 'success',
             "code": 200,
             "error": error
         }
@@ -260,10 +250,57 @@ def kill():
 
     return {
         "path": request.path,
-        "status": 'success' if not error else 'fail',
+        "status": 'success',
         "code": 200,
         "error": error
     }
+
+
+def get_schedule():
+    """Return the current schedule."""
+
+    sem.acquire()
+    report = schedule.get_schedule()
+    sem.release()
+    return {
+        "path": request.path,
+        "status": 'success',
+        "code": 200,
+        "schedule": report,
+        "error": ''
+    }
+
+
+def setup_schedule():
+    """Setup schedule."""
+
+    err = ''
+    sem.acquire()
+    filename = request.json.get('schedule')
+    clear = request.json.get('clear', False)
+    if clear:
+        schedule.clear_schedule()
+    if filename and os.path.exists(filename) and os.path.isfile(filename):
+        err = schedule.read_schedule(filename)
+    sem.release()
+    if err:
+        abort(400, err)
+    return {
+        "path": request.path,
+        "status": 'success',
+        'code': 200,
+        "error": err
+    }
+
+
+def check_schedule():
+    """Check schedule in the background."""
+
+    while True:
+        sem.acquire()
+        schedule.check_schedule()
+        sem.release()
+        gevent.sleep(10)
 
 
 @app.route('/')
@@ -298,8 +335,28 @@ def execute_command(command):
         elif command == 'kill':
             # Results won't make it back if successful
             results = kill()
+        elif command == 'scheduler':
+            results = setup_schedule()
         else:
             abort(404)
+    else:
+        abort(404)
+
+    return results
+
+
+@app.route('%s/scheduler/<string:command>' % get_api_ver_path(), methods=['GET'])
+@auth.login_required
+def get_scheduler(command):
+    """Retrieve information from scheduler."""
+
+    if request.method == 'GET':
+        if command == 'schedule':
+            results = get_schedule()
+        else:
+            abort(404)
+    else:
+        abort(404)
 
     return results
 
@@ -367,19 +424,31 @@ def server_error(error):
     )
 
 
-def run(host=HOST, port=PORT, device_index=0, device_path=None, token=None, ssl=None, debug=False, **kwargs):
+def run(
+    host=HOST, port=PORT, device_index=0, device_path=None, token=None, schedule_file=None,
+    debug=False, **kwargs
+):
     """Run server."""
 
     global luxafor
     global http_server
     global tokens
+    global schedule
 
     with usb.Luxafor(device_index, device_path, token) as lf:
         luxafor = lf
         tokens = set([token])
+        schedule = scheduler.Scheduler(luxafor)
+        if schedule_file and os.path.exists(schedule_file) and os.path.isfile(schedule_file):
+            err = schedule.read_schedule(schedule_file)
+            if err:
+                print(err)
+        http_server = WSGIServer((host, port), app, **kwargs)
+        serve = gevent.spawn(http_server.start)
+        background = gevent.spawn(check_schedule)
 
         try:
-            http_server = WSGIServer((host, port), app, **kwargs)
-            http_server.serve_forever()
+            gevent.joinall([serve, background])
         except KeyboardInterrupt:
             pass
+        print('Exiting')