+++ docs/src/markdown/about/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.2.0
+
+- **NEW**: `--wait` removed from CLI as the server should not get hung while waiting. `wait` is still available when
+  using `pyluxa4.usb` directly.
+- **NEW**: Add new `scheduler` command that allows scheduling command events based on day and time and clearing
+  scheduled events.
+- **NEW**: Add new `get` command which can be used to retrieve currently loaded schedules via `pyluxa4 get schedule`.
+
 ## 1.1.0
 
 - **NEW**: Allow `traffic-light`, `police`, `rainbow`, and `random<1 - 5>` as valid values on CLI along with numerical